requirements.txt
streamlit
python-dotenv
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
pypdf